So fast!
I didn't expect that having the set of "seed boards" (upon which the
other boards are overlayed on top of) be symmetrically unique would
improve the speed so much. I thought it would still be slow since the
algorithm is so brute force but it seems fast enough to just try out a
set of words quick and see if it works. Woa! I bet I could even get this
implementation in elisp and it wouldn't be abysmal either. I'm so happy
and shocked.