Fixed a bug by generalizing
That's always nice! Turns out that I tried to manually optimize things
by hardcoding the initial letter placement so it didn't contain any
symmetries but I pruned too much!

I'm very happy now because I thought I couldn't generate a boggle puzzle
with the boggle group folks that are still going strong (or semi-strong)
but it seems that it is possible! It makes our group feel... destined or
"true" somehow as silly as that sounds since most friend groups in the
world could have their name turned into a boggle puzzle lol. Still, it's
kind of cool that we truly do need to fill up every slot to get a viable
solution.

A couple more stats, if the set of words is:

```
"veronica", "lucas", "danny", "anais", "jane", "joe"
```

Then there are 10 possible solutions. If we add the nicknames "janey"
and "joey" (so the list I input is):

```
"veronica", "lucas", "danny", "anais", "janey", "joey"
```

Then there are only 2 solutions. Woa.